requirements.txt
buttons==0.1.9
dnspython==2.1.0
motor==2.4.0
pip-chill==1.0.1

attrs~=21.2.0
pymongo~=3.11.4
matplotlib~=3.4.2